Move Integration Tests
* Move integration tests to tests/integration.rs
* Update find_unused to search tests direcotory
* Add src/lib.rs to enable the move to tests
