+++ src/lib.rs
@@ -1,0 +1,3 @@
+pub mod http;       // expose your http.rs module
+pub mod output;     // add other modules you want to access in tests
+pub mod cli;        // same idea if needed
